With 128 bit zkeys: all tests passing; pip install (+play_xiangqi) standard and editable both work.

+++ src/core/common/common.hpp
@@ -135,7 +135,7 @@ struct BoardDirection {
 //   }
 // };
 
-typedef uint64_t zkey_t;
+typedef __uint128_t zkey_t;
 typedef array<array<zkey_t, kNumFiles>, kNumRanks> piece_zarray_t;
 typedef array<piece_zarray_t, kNumPieceTypeVals> team_zarray_t;
 typedef array<team_zarray_t, 2> game_zarray_t;
